Use guint instead of ulong for signal ID.
2005-01-31  Hans Petter Jansson  <hpj@novell.com>

	* gui/widgets/eab-config.c (_EABConfigPrivate): Use guint instead of
	ulong for signal ID.

svn path=/trunk/; revision=28641

+++ addressbook/gui/widgets/eab-config.c
@@ -29,7 +29,7 @@ static GObjectClass *ecp_parent_class;
 static GObjectClass *ecph_parent_class;
 
 struct _EABConfigPrivate {
-	ulong source_changed_id;
+	guint source_changed_id;
 };
 
 #define _PRIVATE(o) (g_type_instance_get_private((GTypeInstance *)o, eab_config_get_type()))
